Add keyword arguments (with correct answers).
