+++ src/about_methods.rb
@@ -77,36 +77,6 @@ class AboutMethods < EdgeCase::Koan
     assert_equal __([:one]), method_with_var_args(:one)
     assert_equal __([:one, :two]), method_with_var_args(:one, :two)
   end
- # ------------------------------------------------------------------
-
-  if ruby_version?('2.0')
-    def method_with_keyword_arguments(one: 1, two: 'two')
-      [one, two]
-    end
-
-    def test_keyword_arguments
-      assert_equal __, method_with_keyword_arguments.class
-      assert_equal __, method_with_keyword_arguments
-      assert_equal __, method_with_keyword_arguments(one: 'one')
-      assert_equal __, method_with_keyword_arguments(two: 2)
-    end
-
-    def method_with_keywork_arguments_with_mandatory_argument(one, two: 2, three: 3)
-      [one, two, three]
-    end
-
-    def test_keyword_arguments_with_wrong_number_of_arguments
-      exception = assert_raise (__) do
-        method_with_keywork_arguments_with_mandatory_argument
-      end
-      assert_match(/__/, exception.message)
-    end
-
-    # THINK ABOUT IT:
-    #
-    # Keyword arguments always have a default value, making them optional to the caller
-
-  end
 
   # ------------------------------------------------------------------
