+++ src/edgecase.rb
@@ -65,7 +65,7 @@ class Object
     end
   end
 
-  in_ruby_version("1.9") do
+  in_ruby_version("1.9", "2") do
     public :method_missing
   end
 end
