+++ src/path_to_enlightenment.rb
@@ -12,6 +12,9 @@ require 'about_strings'
 require 'about_symbols'
 require 'about_regular_expressions'
 require 'about_methods'
+in_ruby_version("2") do
+  require 'about_keyword_arguments'
+end
 require 'about_constants'
 require 'about_control_statements'
 require 'about_true_and_false'
